pyproject.toml
[tool.poetry]
name = "fleecekmbackend"
version = "0.1.0"
description = ""
authors = ["nociza <azicon@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
databases = "^0.9.0"
aiosqlite = "^0.20.0"
pandas = "^2.2.1"
uvicorn = "^0.28.0"
requests = "^2.31.0"
together = "^0.2.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
